feat: made CREDEBL white label
