+++ apps/organization/templates/organization-onboard.template.ts
@@ -7,7 +7,7 @@ export class OnBoardVerificationRequest {
       return `<!DOCTYPE html>
             <html lang="en">
               <head>
-                <title>WELCOME TO CREDEBL</title>
+                <title>WELCOME TO ${process.env.PLATFORM_NAME}</title>
                 <meta charset="UTF-8" />
                 <meta name="viewport" content="width=device-width, initial-scale=1" />
               </head>
@@ -16,7 +16,7 @@ export class OnBoardVerificationRequest {
                 <div
                   style="
                     margin: auto;
-                    width: 40%;
+                    max-width: 450px;
                     padding: 20px 30px;
                     background-color: #ffffff;
                     display: block;
@@ -24,8 +24,12 @@ export class OnBoardVerificationRequest {
                 >
                   <div style="display: block; text-align: center">
                     <img
-                      src="${process.env.API_GATEWAY_PROTOCOL}://${process.env.API_ENDPOINT}/CREDEBL_Logo.png"
+                      src="${process.env.API_GATEWAY_PROTOCOL}://${process.env.API_ENDPOINT}/${process.env.PLATFORM_LOGO}"
                       alt="Credebl Logo"
+                      style="max-width:100px"
+                      width="100%"
+                      class="CToWUd"
+                      data-bit="iit"
                     />
                   </div>
                   <div style="display: block; text-align: center">
@@ -33,6 +37,7 @@ export class OnBoardVerificationRequest {
                       src="${process.env.API_GATEWAY_PROTOCOL}://${process.env.API_ENDPOINT}/invite.png"
                       alt="Invite Image"
                       style="max-width: 100%"
+                      class="CToWUd"
                     />
                   </div>
                   <div
@@ -49,7 +54,7 @@ export class OnBoardVerificationRequest {
                     <p>The ${orgName} has been sent an onboard request</p>
                     <p>
                       In case you need any assistance to access your account, please contact
-                      <a href="https://blockster.global" target="_blank">Blockster Labs</a>
+                      <a href=${process.env.POWERED_BY_URL} target="_blank">${process.env.POWERED_BY}</a>
                     </p>
                     <hr style="border-top: 1px solid #e8e8e8" />
                     <footer style="padding-top: 20px">
@@ -72,7 +77,7 @@ export class OnBoardVerificationRequest {
                         /></a>
                       </div>
                       <p style="margin-top: 2px">
-                        &reg; CREDEBL ${year}, Powered by Blockster Labs. All Rights Reserved.
+                        &reg; ${process.env.PLATFORM_NAME} ${year}, Powered by ${process.env.POWERED_BY}. All Rights Reserved.
                       </p>
                     </footer>
                   </div>
