+++ apps/organization/templates/organization-url-template.ts
@@ -25,13 +25,10 @@ export class URLOrganizationEmailTemplate {
             </head>
             
             <body style="margin: 0px; padding:0px; background-color:#F9F9F9;">
-                <div style="margin: auto; width: 40%; padding: 20px 30px; background-color: #FFFFFF; display:block;">
+                <div style="margin: auto; max-width: 450px; padding: 20px 30px; background-color: #FFFFFF; display:block;">
                     <div style="display: block; text-align:center;">
-                        <img src="${process.env.API_GATEWAY_PROTOCOL}://${process.env.API_ENDPOINT}/CREDEBL_Logo.png" alt="Credebl Logo">
+                        <img src="${process.env.API_GATEWAY_PROTOCOL}://${process.env.API_ENDPOINT}/${process.env.PLATFORM_LOGO} alt="Credebl Logo" style="max-width:100px" width="100%" class="CToWUd" data-bit="iit">
                     </div>
-                    <div style="display: block; text-align:center;">
-                        <img src="${process.env.API_GATEWAY_PROTOCOL}://${process.env.API_ENDPOINT}/verification-image.png" alt="verification Image" style="max-width:100%;">
-                        </div>
                     <div style="font-family: Montserrat; font-style: normal; font-weight: 500;
                     font-size: 15px; line-height: 24px;color: #5E5873;">
                         <p style="margin-top:0px">
@@ -53,8 +50,8 @@ export class URLOrganizationEmailTemplate {
                                 VERIFY
                             </a>
                         </div>
-                        <p>In case you need any assistance to access your account, please contact <a href="https://blockster.global"
-                                target="_blank">Blockster Labs</a>
+                        <p>In case you need any assistance to access your account, please contact <a href=${process.env.POWERED_BY_URL}
+                                target="_blank">${process.env.POWERED_BY}</a>
                         </p>
                         <hr style="border-top:1px solid #e8e8e8" />
                         <footer style="padding-top: 20px;">
@@ -66,7 +63,7 @@ export class URLOrganizationEmailTemplate {
                                         width="18" height="18" alt="t" style="color:#cccccc;"></a>
                             </div>
                             <p style="margin-top: 2px;">
-                                &reg; CREDEBL ${year}, Powered by Blockster Labs. All Rights Reserved.
+                                &reg; ${process.env.PLATFORM_NAME} ${year}, Powered by ${process.env.POWERED_BY}. All Rights Reserved.
                             </p>
                         </footer>
                     </div>
